Move date formatting code to utils file
