+++ frontend/src/metabase/components/DateQuarterYearWidget/DateQuarterYearWidget.tsx
@@ -5,6 +5,7 @@ import cx from "classnames";
 import { t } from "ttag";
 
 import YearPicker from "metabase/components/YearPicker";
+import { formatQuarterYearWidget } from "metabase/parameters/utils/date-formatting";
 
 // translator: this is a "moment" format string (https://momentjs.com/docs/#/displaying/format/) It should include "Q" for the quarter number, and raw text can be escaped by brackets. For eample "[Quarter] Q" will be rendered as "Quarter 1" etc
 const QUARTER_FORMAT_STRING = t`[Q]Q`;
@@ -43,10 +44,7 @@ class DateQuarterYearWidget extends React.Component<Props, State> {
     }
   }
 
-  static format = (value: string) => {
-    const m = moment(value, "[Q]Q-YYYY");
-    return m.isValid() ? m.format("[Q]Q, YYYY") : "";
-  };
+  static format = formatQuarterYearWidget;
 
   componentWillUnmount() {
     const { quarter, year } = this.state;
