+++ frontend/src/metabase/components/DateRangeWidget/DateRangeWidget.tsx
@@ -2,13 +2,7 @@ import React from "react";
 
 import moment from "moment";
 import DateAllOptionsWidget from "metabase/components/DateAllOptionsWidget";
-
-const SEPARATOR = "~"; // URL-safe
-
-function parseDateRangeValue(value: string) {
-  const [start, end] = (value || "").split(SEPARATOR);
-  return { start, end };
-}
+import { formatRangeWidget } from "metabase/parameters/utils/date-formatting";
 
 interface DateRangeWidgetProps {
   setValue: (value: string | null) => void;
@@ -31,13 +25,6 @@ const DateRangeWidget = ({ value, ...props }: DateRangeWidgetProps) => {
   );
 };
 
-DateRangeWidget.format = (value: string) => {
-  const { start, end } = parseDateRangeValue(value);
-  return start && end
-    ? moment(start).format("MMMM D, YYYY") +
-        " - " +
-        moment(end).format("MMMM D, YYYY")
-    : "";
-};
+DateRangeWidget.format = formatRangeWidget;
 
 export default DateRangeWidget;
