+++ frontend/src/metabase/components/DateSingleWidget/DateSingleWidget.tsx
@@ -2,6 +2,7 @@ import React from "react";
 
 import moment from "moment";
 import DateAllOptionsWidget from "metabase/components/DateAllOptionsWidget";
+import { formatSingleWidget } from "metabase/parameters/utils/date-formatting";
 
 interface DateSingleWidgetProps {
   setValue: (value: string | null) => void;
@@ -20,7 +21,6 @@ const DateSingleWidget = ({ value, ...props }: DateSingleWidgetProps) => {
   );
 };
 
-DateSingleWidget.format = (value: string) =>
-  value ? moment(value).format("MMMM D, YYYY") : "";
+DateSingleWidget.format = formatSingleWidget;
 
 export default DateSingleWidget;
